log_event: structured event logs
The documentation in this commit is probably what you want to read, but:
this adds ->log_event and ->log_debug_event methods to log structured
data in a form of the compact, human-readable-ish logfmt format, with
automatic expansion of nested structures.

It adds "proxy_ctx" for data to be logged on every line from a proxy
logger, acting like the free-form "proxy_prefix" argument, but with
logfmt data.