event logging: cache proxy context kv strings
If, on a proxy, you have the context data [ x => $complex ], then
computing the string(s) to put into each log message within that proxy
context could be expensive -- or at least more expensive than it has to
be.  Also, since we want to have code refs for lazily-computed values,
like the rest of Log::Dispatchouli has, we want to only call those subs
once.

Fortunately for us, those subs will be a useful way to test this
behavior!

Instead of only storing the raw kv pairs, we'll also cache the computed
kvstrs (the "k=v" strings) once we have to compute them.  This adds one
new sub call, _log_event, to every event log.  On the other hand, it
eliminates the format_event call.  This may be a loss for future
subclassing but… we can sort that out when we get there.