tests: improve naming to be less confusing
Wow, what was I thinking?

I had a setup like this:

    $logger
      └─ proxy1
          └─ proxy2

...and the two proxies had a prefix.  "Clearly", proxy2 is more interior
than proxy1, because we create it as we descend into a structure
(probably a loop), and proxy1 is out toward the top, outermost level.

Despite that, I set the prefix of proxy2 to "outer" and proxy1 to
"inner".  This led me to confuse the ordering of the prefixes and think
it was a bug.  It wasn't a bug, it worked.  I just chose terrible prefix
strings for these tests.

This fixes it.