introduce flog_messages and fmt_event
These methods return the string that would be logged without actually
logging it.  The logging methods are then rewritten in terms of this.

I'll be using this to get our proxy prefix, proxy context, and
formatting behavior introduced to a piece of software that needs to log
to both syslog *and* a separate log with different content, but using
the same context.